DRY the code by creating the Sushi::AttributePerformance module. now using the above module in the platform, platform usage and item reports.
